REF: make pd._testing a directory (#38689)
